Extended the wiringPiISR() function
It is often useful, if the function that gets called to handle an occurred interrupt, can be passed
user defined data, such as a pointer to a data structure or a counter.
Since the ISR is internally handled by a pthread that provides an void* pointer as argument to the
created pthread function anyway - this adds almost no overhead.
Example given in examples/isr.c: wiringPiISR is called on 8 pins and every call uses the same
function but a different pointer to an integer counter is provided.